made ReleaseSort component
