+++ components/Profile/Profile.jsx
@@ -5,26 +5,27 @@ import SocialSites from "../SocialSites/SocialSites"
 import { useState } from "react"
 import Head from "next/head"
 import Link from "next/link"
+import ReleaseSort from "../ReleaseSort/ReleaseSort"
 
-const sortByCreatedAt = (a, b) => (a.created_at > b.created_at ? 1 : -1)
-const sortByTitle = (a, b) => (a.title > b.title ? 1 : -1)
+// const sortByCreatedAt = (a, b) => (a.created_at > b.created_at ? 1 : -1)
+// const sortByTitle = (a, b) => (a.title > b.title ? 1 : -1)
 
-const filterOptions = [
-  {
-    label: "Oldest First",
-    value: "oldest",
-    method: sortByCreatedAt,
-    direction: "asc",
-  },
-  {
-    label: "Newest First",
-    value: "newest",
-    method: sortByCreatedAt,
-    direction: "desc",
-  },
-  { label: "A to Z", value: "a-z", method: sortByTitle, direction: "asc" },
-  { label: "Z to A", value: "z-a", method: sortByTitle, direction: "desc" },
-]
+// const filterOptions = [
+//   {
+//     label: "Oldest First",
+//     value: "oldest",
+//     method: sortByCreatedAt,
+//     direction: "asc",
+//   },
+//   {
+//     label: "Newest First",
+//     value: "newest",
+//     method: sortByCreatedAt,
+//     direction: "desc",
+//   },
+//   { label: "A to Z", value: "a-z", method: sortByTitle, direction: "asc" },
+//   { label: "Z to A", value: "z-a", method: sortByTitle, direction: "desc" },
+// ]
 
 export default function ProfileLayout({
   avatar,
@@ -46,19 +47,6 @@ export default function ProfileLayout({
   )
   const [showError, setShowError] = useState(false)
   const [sortedReleases, setSortedReleases] = useState(releases)
-
-  const handleSort = (value) => {
-    const selected = filterOptions.find((option) => option.value === value)
-    let sortedItems
-
-    sortedItems = [...releases].sort(selected.method)
-
-    if (selected.direction === "desc") {
-      sortedItems.reverse()
-    }
-
-    setSortedReleases(sortedItems)
-  }
 
   function handleSubmit(e) {
     e.preventDefault()
@@ -152,23 +140,10 @@ export default function ProfileLayout({
         ) : (
           <>
             {isSubscribed || isDlcmFriend ? (
-              <div className={styles.filter}>
-                <label className="label" htmlFor="order">
-                  Order
-                </label>
-                <select
-                  className="input select"
-                  style={{ inlineSize: "auto" }}
-                  id="order"
-                  onChange={(e) => handleSort(e.target.value)}
-                >
-                  {filterOptions.map(({ label, value }) => (
-                    <option value={value} key={value}>
-                      {label}
-                    </option>
-                  ))}
-                </select>
-              </div>
+              <ReleaseSort
+                releases={releases}
+                setSortedReleases={setSortedReleases}
+              />
             ) : null}
             <ul className="grid" role="list">
               {sortedReleases.map((release, index) =>
