+++ components/ReleaseSort/ReleaseSort.module.css
@@ -1,0 +1,8 @@
+.filter {
+  display: flex;
+  align-items: center;
+}
+
+.filter select {
+  margin-inline-start: 1em;
+}
